can: c_can: use proper type for 'instance'
[ Upstream commit f323d7a1d2868c00b2604dca36ad82e8ecbe4270 ]

Commit 6439fbce1075 (can: c_can: fix error checking of priv->instance in
probe()) found the warning but applied a suboptimal solution. Since, both
pdev->id and of_alias_get_id() return integers, it makes sense to convert the
variable to an integer and avoid the cast.

+++ drivers/net/can/c_can/c_can_platform.c
@@ -227,7 +227,7 @@ static int c_can_plat_probe(struct platform_device *pdev)
 							     res->start,
 							     resource_size(res));
 
-		if (!priv->raminit_ctrlreg || (int)priv->instance < 0)
+		if (!priv->raminit_ctrlreg || priv->instance < 0)
 			dev_info(&pdev->dev, "control memory is not used for raminit\n");
 		else
 			priv->raminit = c_can_hw_raminit;
